Flask sites should always set SECRET_KEY
Even if we know a site doesn't use it, we should set this key
as it's very easy for someone to come along later and add features
that rely on it without remembering to set up the secret key.

We should then add a default insecure secret key setting into `.env`
in our projects.