Check static files against provided hashes
When requesting a static file under a `/static` URL,
a `?v=` parameter may be provided. This should contain
the beginning of an MD5 hex digest.

If this parameter is present, FlaskBase will check if this
MD5 prefix matches the MD5 of the contents of the file being
requested. If not it will return a 404. If yes, it will
extend the cacheable time of the file to 1 year.

Tests are included to check this functionality.