Fix cache-control "none" test, convert to "zero"
In retrospect, the test of setting cache-control values to "None"
was always going to be confusing and liable to product errors.

And so it is: https://github.com/canonical-web-and-design/canonicalwebteam.flask-base/issues/48

If you want to explicitly set cache-control values for a view,
using None is a very bad idea. They should always be set to an int
instead.

So I'm changing the test to test setting these values to 0 explicitly.